More casting

+++ DBlib/src/DBImpASCIINet.C
@@ -143,7 +143,7 @@ int DBImportASCIINet(DBObjData *netData, const char *fileName) {
                     subbasinLengthFLD->Float(cellRec, (DBFloat) 0.0);
                     subbasinAreaFLD->Float(cellRec, (DBFloat) 0.0);
 
-                    ((DBInt *) dataRec->Data())[pos.Row * colNum + pos.Col] = cellRec->RowID();
+                    ((DBInt *) dataRec->Data())[(size_t) pos.Row * (size_t) colNum + (size_t) pos.Col] = cellRec->RowID();
                 }
             }
         }
